Made component tree shakable
