+++ src/lib/components/Asset/index.ts
@@ -1,0 +1,1 @@
+export * from "./AssetWrapper";
