+++ src/lib/config.ts
@@ -1,5 +1,5 @@
 type Env = "dev" | "prod";
-type Config = {
+export type Config = {
   authKey: string;
   env?: Env;
   cache?: boolean;
@@ -10,9 +10,3 @@ export const config: Config = {
   env: "dev",
   cache: true,
 };
-
-export function init(key: string, _config?: Omit<Config, "authKey">) {
-  config.authKey = key;
-  config.env = _config?.env || "dev";
-  config.cache = _config?.cache === false ? false : true;
-}
