+++ src/lib/hooks/index.ts
@@ -1,0 +1,14 @@
+import { useQuery, useLazyQuery } from "./useQuery";
+import {
+  useQueryWithPagination,
+  useLazyQueryWithPagination,
+} from "./useQueryWithPagination";
+
+export {
+  useQuery,
+  useQueryWithPagination,
+  useLazyQuery,
+  useLazyQueryWithPagination,
+};
+
+export * from "./popular-queries";
