+++ src/lib/index.ts
@@ -1,22 +1,4 @@
-import { fetchQuery } from "./apis/fetchQuery";
-import { useQuery, useLazyQuery } from "./hooks/useQuery";
-import {
-  useQueryWithPagination,
-  useLazyQueryWithPagination,
-} from "./hooks/useQueryWithPagination";
-import { Asset } from "./components/NFTAsset/AssetWrapper";
-import { init } from "./config";
-import { fetchQueryWithPagination } from "./apis/fetchQueryWithPagination";
-
-export {
-  init,
-  useQuery,
-  useQueryWithPagination,
-  useLazyQuery,
-  useLazyQueryWithPagination,
-  fetchQuery,
-  fetchQueryWithPagination,
-  Asset,
-};
-
-export * from "./hooks/popular-queries";
+export * from "./apis";
+export * from "./init";
+export * from "./hooks";
+export * from "./components";
